footer logo and email varification updated

+++ accounts/models.py
@@ -5,9 +5,9 @@ from django.contrib.auth.models import User
 class UserAccount(models.Model):
     user = models.OneToOneField(User, related_name='account', on_delete=models.CASCADE)
     ssc_roll = models.IntegerField(null=True, blank=True)
-    ssc_gpa = models.IntegerField(null=True, blank=True)
+    ssc_gpa = models.DecimalField(max_digits=5, decimal_places=2,null=True, blank=True)
     hsc_roll = models.IntegerField(null=True, blank=True)
-    hsc_gpa = models.IntegerField(null=True, blank=True)
+    hsc_gpa = models.DecimalField(max_digits=5, decimal_places=2,null=True, blank=True)
 
 
     def __str__(self):
